refactor: use instruction text instead of button
Since the game is played with the keyboard or gamepad, it doesn't make
sense to require using a mouse to start the game. Instead have a message
prompt to tell the user to press the Z or A Button to start the game or
press the ESC or B Button to quit the game.